move findRole into handler- only used there now
